+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/ContributionBindsFunctionsIrTransformer.kt
@@ -152,9 +152,8 @@ internal class ContributionTransformer(private val context: IrMetroContext) :
     // Add fake overrides. This should only add missing ones
     declaration.addFakeOverrides(irTypeSystemContext)
     if (!declaration.isAnnotatedWithAny(metroSymbols.classIds.graphExtensionAnnotations)) {
-      // Only DependencyGraph classes have an FIR-generated graph impl. ContributesGraphExtension
-      // will get
-      // implemented later in IR
+      // Only DependencyGraph classes have an FIR-generated graph impl. Contributed GraphExtensions
+      // will get implemented later in IR
       declaration
         .requireNestedClass(Origins.GraphImplClassDeclaration)
         .addFakeOverrides(irTypeSystemContext)